+++ packages/antwerp-ui/react-components/src/lib/atoms/table/Table.types.ts
@@ -5,6 +5,7 @@ export interface TableProps {
   actions?: {
     label?: string;
     buttons: (ButtonProps & {
+      label?: React.ReactNode;
       onActionClick?: (event: MouseEvent<HTMLButtonElement>, rowId: string) => void;
     })[];
   };