+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 7.0.1
+Version 7.0.2
 
 # Changelog
 
@@ -7,9 +7,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 7.0.2 - 2023-08-30
+
 =======
+
 ## Added
+
 - Can use both the `state` and `description` property for the `Autocomplete` component
 - Can use state and description in the `inputProps` for the Datepicker component